+++ NEWS.md
@@ -54,7 +54,7 @@ Note: We're only listing outstanding class updates.
     * Notify the Ruby virtual machine that the boot sequence is finished,
       and that now is a good time to optimize the application. This is useful
       for long running applications. The actual optimizations performed are entirely
-      implementation specific and may change in the future without notice. [[Feature #18885]
+      implementation specific and may change in the future without notice. [[Feature #18885]]
 
 * Refinement